--- App/Scene/Login/styles.js
@@ -1,7 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
